requirements-dev.txt
-r requirements.txt
boto3-stubs[lambda,sqs,events]
pytest==7.2.1
pytest-cov==4.0.0
pytest-mock==3.5.1
pytest-dotenv==0.5.2
flake8==3.9.2
fastapi==0.65.2
uvicorn==0.18.1
moto==2.2.14
pip-licenses==3.5.3
git+ssh://git@github.com/jitsecurity/test-utils.git@1.1.3#egg=test-utils
mypy==0.961
pydantic-factories==1.1.0
types-requests
mongomock==4.1.2
